Fixed Root Sidebar Active Class Bug
Note: it isn't actually fixed yet, but the change was necessary for us to fix it later.

It will be actually fixed when the ".html" at the end of the links are removed and the "/index.html" is replaced with "/", this will happen before we start the add responsiveness for mobile users.