Added Default Active Class to Sidebar
This was a revert of the previous temporary commit named "TEMP - Removed Active Class from Sidebar". This commit makes sure that if the user opens up a url in our website, the correct sidebar button is active by default.

Note: This doesn't break the page switching in any way currently known.